+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Head
+
+* Added: `font-family-name-quotes` rule.
 * Added: better handling of vendor prefixes in `property-unit-blacklist` and `property-unit-whitelist`, e.g. if you enter `animation` it now also checks `-webkit-animation`.
 * Added: support for regular expression property identification in `property-blacklist`, `property-unit-blacklist`, `property-unit-whitelist`, `property-value-blacklist`, and `property-whitelist`.
 * Added: brought `stylelint-rule-tester` into this repo, and exposed it at `stylelint.utils.ruleTester`.